refactor(ParsingService): Service 객체를 무상태로 유지하도록 변경
