+++ src/main/java/calculator/Application.java
@@ -4,7 +4,6 @@ import calculator.controller.CalculatorController;
 
 public class Application {
     public static void main(String[] args) {
-        CalculatorController calculatorController = new CalculatorController();
         CalculatorController.run();
     }
 }
